Rolling 24h agent limit, invariant tests, CI, threat model; remove voice input
- AgentVault v2: the daily cap is a strict rolling 24-hour window (ring buffer of the
  last 32 spends per token), so there is no midnight double-spend. version() = 2.
  Factory rejects a zero router. Not deployed yet: v1 factories stay configured.
- Foundry invariant test: no 24-hour period ever exceeds the cap (128k random calls
  with time jumps). It fails against v1 (101.97 > 100), which is how the bug was found.
- Agent server safety checks moved to pure server/guards.ts with unit tests:
  recipients only from the owner's latest message, model-suggested .arc names rejected,
  per-tx / 24h / mainnet beta caps, and a rolling guard for existing v1 vaults.
- Vitest unit tests for validateSend and intentParser; `npm test`.
- CI (GitHub Actions): typecheck, type-aware lint, unit tests, build, forge tests.
  Fixed the existing lint errors in src/ and server/ so the lint step is green.
- SECURITY.md: trust boundaries, worst case per trusted component, Slither results,
  verified-source links (mainnet on Sourcify, testnet on the Arc explorer).
- Agent panel finds vaults from legacy factories and suggests upgrading them.
- Removed voice input (mic button, speech hook, normaliser, docs).

+++ src/hooks/useTokenBalances.ts
@@ -22,7 +22,7 @@ export function useTokenBalances(account: `0x${string}` | undefined) {
   const balances: TokenBalances = {};
   tokens.forEach((t, i) => {
     const r = data?.[i];
-    if (r?.status === 'success') balances[t.symbol] = r.result as bigint;
+    if (r?.status === 'success') balances[t.symbol] = r.result;
   });
   return { balances, refetch, isError };
 }
